Generate java code from cut down wsdl
This commit adds a cut down version of the wsdl that describes
the DRS web service. It only contains a single operation (checkAvailability).
We did do a lot of manual XML editing to get to this version.
This new wsdl successfully generates Java code but we need to get it to
run a server before we are confident it works correctly.

We cut it down to a single operation so we can get to a single slice of
functionality quicker.